Massive bug fixes to OpenBLAS solver functions.
Except for gesv the solver functions were pretty much all broken. I
fixed it and added some notes on their design. There is an issue that
gesv and sysv are not mutable while getrf and getri are. I've made sysv
non mutable because I noticed that it was not being backed up.

Except for gesv all of them needed transposition steps at the beginning
and at the end.