Add OpenBLAS@0.3.5 for macOS
Fixes #24.

Ideally this would be the same version as the others (0.2.15) or the
other versions would be updated, but it seems hard to find specific
versions of the library online. The recent "published" releases are hosted on
SourceForge (known lately for malware) and don't actually seem to
include the binaries.

However, this executes the tests correctly on macOS, and clearly the old
version was working for Windows, so this does not break anything
further.