Enable building for macOS (#50)
* Fix incorrectly-placed UTF-16 BOM

The recently-inserted license header (f88e387f) was added before the BOM
(U+FEFF) in some files, causing invalid encodings and a confusing error
when attempting to build the solution, at least on macOS:
> error FS0010: Unexpected character ''

* Add OpenBLAS@0.3.5 for macOS

Fixes #24.

Ideally this would be the same version as the others (0.2.15) or the
other versions would be updated, but it seems hard to find specific
versions of the library online. The recent "published" releases are hosted on
SourceForge (known lately for malware) and don't actually seem to
include the binaries.

However, this executes the tests correctly on macOS, and clearly the old
version was working for Windows, so this does not break anything
further.

* Build for Any CPU

This allows the solution to build on macOS. Fixes #22, I believe.