Document that libtorch-{version} may need to be changed.
The documentation NativeLibrary.Load paths referred to outdated libtorch versions. This needed to be updated. It does not look like there is a {{fsdocs-content}} tag that will work to dynamically insert the proprer libtorch reference. So instead I updated the libtorch version and added a comment that this may need to be changed.